requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0
chromadb>=0.5.0
anthropic>=0.30.0
PyJWT>=2.8.0
bcrypt>=4.0.0
pypdf>=4.0.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
langchain-text-splitters>=0.3.0
